Change return type of MediaPlayerWidget::position() to double.

+++ src/ui/mediaplayerwidget.h
@@ -36,7 +36,7 @@ public:
     ~MediaPlayerWidget();
 
     int duration() const;
-    int position() const;
+    double position() const;
 
 signals:
     void stateChanged();
